requirements.txt
asgiref == 3.7.2       
beautifulsoup4 == 4.12.2      
cffi == 1.16.0      
cryptography == 41.0.4      
cx-Oracle == 8.3.0       
Django == 4.1.2       
django-bootstrap-v5 == 1.0.11      
django-cors-headers == 4.3.0       
djangorestframework ==  3.14.0      
djangorestframework-jwt == 1.11.0      
djangorestframework-simplejwt == 5.3.0       
drf-yasg == 1.21.7
humanize == 4.7.0
inflection == 0.5.1
packaging == 23.2
Pillow == 10.0.0
pip == 23.2.1
pycparser == 2.21
PyJWT == 1.7.1
pytz == 2023.3.post1
PyYAML == 6.0.1
setuptools == 68.2.2
soupsieve == 2.5
sqlparse == 0.4.4
tzdata == 2023.3
uritemplate == 4.1.1
django-corse-headers == 4.3.0
drf-yasg == 1.21.7